% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievementInput.R
\name{achievementInput}
\alias{achievementInput}
\title{Input with achievement description}
\usage{
achievementInput(id, desc = "", topics = "", values = "")
}
\description{
Input with achievement description
}
