% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_elem}
\alias{add_elem}
\title{Add alement to markdown file}
\usage{
add_elem(elem, nbefore = 1L, nafter = 1L, symbefore = "", symafter = "",
  file = cv_file)
}
\description{
Add alement to markdown file
}
